project organization and env parser work
