+++ cmd/hulak/initialize.go
@@ -1,0 +1,20 @@
+package main
+
+import (
+	"github.com/xaaha/hulak/pkg/envparser"
+)
+
+/*
+InitializeProject() starts the project by creating envfolder and global file in it.
+TBC...
+*/
+func InitializeProject() {
+	err := envparser.CreateDefaultEnvs(nil)
+	if err != nil {
+		panic(err)
+	}
+	err = envparser.SetDefaultEnv()
+	if err != nil {
+		panic(err)
+	}
+}
